GFX::FrameDisposable: changes miss counter to serve current GC better

+++ game/gfx/FrameDisposable.h
@@ -17,7 +17,8 @@ class FrameDisposable {
     void decreaseMiss();
 
   private:
-    uint8_t misses = 0;
+    // EVAL concept
+    uint16_t misses = 0;
 };
 
 }
